redid the reading from file part yet again...
Turns out, I'm a right noooooob.
The biggest change now, is this: read_birb() doesn't create a duplicate
buffer anymore, when it translates the read data from u8 to u64. Instead
it reuses the same buffer via Vec::from_raw_parts.

That also brought with it the necessity for some active memory checking,
as we can't translate a u8 Vector, that doesn't contain exactly a
multiple of 8 numbers and has a capacity of exactly a multiple of 8 to a
u64 Vector **safely**. If length and capacity don't match up exactly,
there might be parts of the source vector, that get lost -- permanenly.
This would confuse the allocator.