Made the Iterator Infra-structure for Metropolis-Hastings.
Added following types:
  - Complex	for representing complex numbers
  - Orbit	an iterator that repeatedly applies the mandelbrot
		equation to a seed value
  - MHOrbits	iterator, that yields a number of such
		mandelbrot-orbits, the starting points of which are
		chosen using a Metropolis-Hastings algorithm.

Note: A lot of tiny functions were added, including the actual
mandelbrot function and stuff for M-H, but none of these were already
implemented. This commit only adds the iterator structure I designed,
which makes it easy to read and understand the algorithm, and in the next
step add the actual logic, that does do things!