Integrated the args parser macro and made the args_parser() function.
We can now linearly parse command line arguments! This is amazing. I'm
honestly shocked, that there don't seem to be any unconsidered bugs in
this thing. (But then again, I'm amazed every time I add a thing to this
program, cause every single time I add something I put a ludicrous
amount of design into it aforehand -- something we should be doing all
the time, but c'mon, who does that? ... Truth: It's easy enough to write
good code. It just takes a lot, A LOT of careful planning, for which you
either don't have the want, or the time, usually.)