Implemented v1 of the logging function.
It outputs all the stuff! =)

As with everything else, I went about designing this thing to the end,
before starting to write up the code for it. That was all fine and
dandy, except the logic, that would make the logging function terminate,
when all the threads were done, was bust ;) ... So that was some right
ripe fun. It all works now, should still make it a lil' bit prettier I
think, but otherwise neat.