I'm increasingly confident that it's correct.
I've done more tests and it would seem we are getting there.
The final arbiter of all this will be birb2img.

Found a curious bug though: I can't use Vec::from_raw_parts() to
write the birb vector, or it will get corrupted. Somehow the first
of the numbers gets overridden to something bollocks. I think that
might be to do with the header data of the new Vec being writtin, who
knows... I took my initial method of using u8 pointers and offsetting
them to reconstruct the data structure as u8. That seems to work just
fine.